Added 2 scenes - Stats and Store.
Clicking on a button on Stats will go to Store, and from the Store, you can go back to the stats. This is done with separate scenes and with Navigator C# script (renamed from MainMenu). This will be the primary file to select the scenes.

For the character stats screen, it is still in its preliminary state, and clicking from Soldier #1 to Soldier #5 icons will only take you to the one available Stats screen.