Soldier List is the one that you get initially. It will pull the list/name of 5 soldiers that are currently in your squad/use.
Clicking on one of the names, by default, should try to match that soldier's ID/name with the database and in turn, import a bunch of data that are relevant to the soldier.
For this instance, clicking on "Catherine", by design, *should* load relevant data about her. But for now, I simply used a local array to test out the text/dialogue boxes.