Armor and Weapon can only be purchased once (assuming that you don't have it in your inventory).
Healthpacks are an exception, which you can buy up to 3 of each.... in theory. It's still a little glitchy, and while there are no compile errors, it's just a bug that the InventoryCount only goes up by 1.

I think I will need to put some stuff from Awake() to Update() but not sure which.